ConfigRawParamsTree: merge w/ ConfigRawParams and del
