+++ MissionPlanner.csproj
@@ -260,9 +260,6 @@
     <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParams.uk.resx">
       <DependentUpon>ConfigRawParams.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.uk.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
     <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigREPL.uk.resx">
       <DependentUpon>ConfigREPL.cs</DependentUpon>
     </EmbeddedResource>
@@ -1496,9 +1493,6 @@
     <Compile Include="GCSViews\ConfigurationView\ConfigRawParams.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="GCSViews\ConfigurationView\ConfigRawParamsTree.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
     <Compile Include="GCSViews\ConfigurationView\ConfigSimplePids.cs">
       <SubType>UserControl</SubType>
     </Compile>
@@ -1836,9 +1830,6 @@
     </Compile>
     <Compile Include="GCSViews\ConfigurationView\ConfigRawParams.Designer.cs">
       <DependentUpon>ConfigRawParams.cs</DependentUpon>
-    </Compile>
-    <Compile Include="GCSViews\ConfigurationView\ConfigRawParamsTree.Designer.cs">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
     </Compile>
     <Compile Include="GCSViews\ConfigurationView\ConfigSimplePids.Designer.cs">
       <DependentUpon>ConfigSimplePids.cs</DependentUpon>
@@ -4322,51 +4313,6 @@
     </EmbeddedResource>
     <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParams.zh-TW.resx">
       <DependentUpon>ConfigRawParams.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.ar.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.az-Latn-AZ.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.de-DE.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.fr.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.id-ID.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.it-IT.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.ja-JP.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.ko-KR.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.pt.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.ru-KZ.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.tr.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.zh-Hans.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.zh-Hant.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigRawParamsTree.zh-TW.resx">
-      <DependentUpon>ConfigRawParamsTree.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="GCSViews\ConfigurationView\ConfigSimplePids.resx">
       <DependentUpon>ConfigSimplePids.cs</DependentUpon>
